PPE/requirements.txt
streamlit>=1.0.0
ultralytics==8.3.34
opencv-python==4.10.0.84
numpy==1.26.4
torch==2.3.1+cpu
torchvision==0.18.1+cpu
mediapipe==0.10.14
Pillow==10.4.0
PyYAML==6.0.2
protobuf==5.27.3
typing-extensions==4.12.2
requests>=2.31.0
scipy==1.11.4
python-dotenv>=1.0.0
